+++ cpp/src/examples/cython_test.cpp
@@ -3,6 +3,8 @@
 //
 #include "lib/Circle.h"
 #include "iostream"
+#include "data/table_builder.h"
+#include "status.cpp"
 
 using namespace std;
 
@@ -15,6 +17,9 @@ int main(){
     cout << "Circle Area (Int) : " << circle.getArea() << endl;
 
 
+    twisterx::Status  s = twisterx::data::read_csv("s", "s");
+
+    cout << s.get_code() << endl;
