+++ cpp/src/twisterx/arrow/arrow_kernels.hpp
@@ -7,205 +7,210 @@
 
 namespace twisterx {
 
-class ArrowPartitionKernel {
-  /**
-   * We partition the table and return the indexes as an array
-   * @param ctx
-   * @param values
-   * @param out
-   * @return
-   */
-  virtual int Partition(arrow::compute::FunctionContext* ctx, const arrow::Table& values,
-                        arrow::Int32Array* out) = 0;
-};
+    class ArrowPartitionKernel {
+        /**
+         * We partition the table and return the indexes as an array
+         * @param ctx
+         * @param values
+         * @param out
+         * @return
+         */
+        virtual int Partition(arrow::compute::FunctionContext *ctx, const arrow::Table &values,
+                              arrow::Int32Array *out) = 0;
+    };
 
-class ArrowTableMergeKernel {
-public:
-  explicit ArrowTableMergeKernel() {}
+    class ArrowTableMergeKernel {
+    public:
+        explicit ArrowTableMergeKernel() {}
 
-  /**
-   * Merge the values in the column and return an array
-   * @param ctx
-   * @param values
-   * @param targets
-   * @param out_length
-   * @param out
-   * @return
-   */
-  virtual int Merge(arrow::compute::FunctionContext* ctx, const arrow::Table& values,
-                        const arrow::Int32Array& partitions,
-                        std::unordered_map<int, std::shared_ptr<arrow::Table>>& out) = 0;
-};
+        /**
+         * Merge the values in the column and return an array
+         * @param ctx
+         * @param values
+         * @param targets
+         * @param out_length
+         * @param out
+         * @return
+         */
+        virtual int Merge(arrow::compute::FunctionContext *ctx, const arrow::Table &values,
+                          const arrow::Int32Array &partitions,
+                          std::unordered_map<int, std::shared_ptr<arrow::Table>> &out) = 0;
+    };
 
-class ArrowArrayMergeKernel {
-public:
-  explicit ArrowArrayMergeKernel(std::shared_ptr<arrow::DataType> type,
-      arrow::MemoryPool* pool, std::shared_ptr<std::vector<int>> targets) : type_(type), pool_(pool),
-      targets_(targets) {}
+    class ArrowArrayMergeKernel {
+    public:
+        explicit ArrowArrayMergeKernel(std::shared_ptr<arrow::DataType> type,
+                                       arrow::MemoryPool *pool, std::shared_ptr<std::vector<int>> targets) : type_(
+                type), pool_(pool),
+                                                                                                             targets_(
+                                                                                                                     targets) {}
 
-  /**
-   * Merge the values in the column and return an array
-   * @param ctx
-   * @param values
-   * @param targets
-   * @param out_length
-   * @param out
-   * @return
-   */
-  virtual int Merge(std::shared_ptr<arrow::Array> &values,
-                    std::shared_ptr <arrow::Int32Array>& partitions,
-                    std::unordered_map<int, std::shared_ptr<arrow::Array>>& out) = 0;
-protected:
-  std::shared_ptr<arrow::DataType> type_;
-  arrow::MemoryPool* pool_;
-  std::shared_ptr<std::vector<int>> targets_;
-};
+        /**
+         * Merge the values in the column and return an array
+         * @param ctx
+         * @param values
+         * @param targets
+         * @param out_length
+         * @param out
+         * @return
+         */
+        virtual int Merge(std::shared_ptr<arrow::Array> &values,
+                          std::shared_ptr<arrow::Int32Array> &partitions,
+                          std::unordered_map<int, std::shared_ptr<arrow::Array>> &out) = 0;
 
-template <typename TYPE>
-class ArrowArrayNumericMergeKernel : public ArrowArrayMergeKernel {
-public:
-  explicit ArrowArrayNumericMergeKernel(std::shared_ptr<arrow::DataType> type,
-                                 arrow::MemoryPool* pool, std::shared_ptr<std::vector<int>> targets) :
-                                 ArrowArrayMergeKernel(type, pool, targets) {}
+    protected:
+        std::shared_ptr<arrow::DataType> type_;
+        arrow::MemoryPool *pool_;
+        std::shared_ptr<std::vector<int>> targets_;
+    };
 
-  int Merge(std::shared_ptr<arrow::Array> &values,
-      std::shared_ptr <arrow::Int32Array>& partitions,
-                    std::unordered_map<int, std::shared_ptr<arrow::Array>>& out) override {
-    auto reader =
-        std::static_pointer_cast<arrow::NumericArray<TYPE>>(values);
-    std::unordered_map<int, std::shared_ptr<arrow::NumericBuilder<TYPE>>> builders;
+    template<typename TYPE>
+    class ArrowArrayNumericMergeKernel : public ArrowArrayMergeKernel {
+    public:
+        explicit ArrowArrayNumericMergeKernel(std::shared_ptr<arrow::DataType> type,
+                                              arrow::MemoryPool *pool, std::shared_ptr<std::vector<int>> targets) :
+                ArrowArrayMergeKernel(type, pool, targets) {}
 
-    for (auto it = targets_.get()->begin() ; it != targets_.get()->end(); ++it) {
-      std::shared_ptr<arrow::NumericBuilder<TYPE>> b = std::make_shared<arrow::NumericBuilder<TYPE>>(type_, pool_);
-      builders.insert(std::pair<int, std::shared_ptr<arrow::NumericBuilder<TYPE>>>(*it, b));
-    }
+        int Merge(std::shared_ptr<arrow::Array> &values,
+                  std::shared_ptr<arrow::Int32Array> &partitions,
+                  std::unordered_map<int, std::shared_ptr<arrow::Array>> &out) override {
+            auto reader =
+                    std::static_pointer_cast<arrow::NumericArray<TYPE>>(values);
+            std::unordered_map<int, std::shared_ptr<arrow::NumericBuilder<TYPE>>> builders;
 
-    for (int64_t i = 0; i < partitions->length(); i++) {
-      std::shared_ptr<arrow::NumericBuilder<TYPE>> b = builders[partitions->Value(i)];
-      b->Append(reader->Value(i));
-    }
+            for (auto it = targets_.get()->begin(); it != targets_.get()->end(); ++it) {
+                std::shared_ptr<arrow::NumericBuilder<TYPE>> b = std::make_shared<arrow::NumericBuilder<TYPE>>(type_,
+                                                                                                               pool_);
+                builders.insert(std::pair<int, std::shared_ptr<arrow::NumericBuilder<TYPE>>>(*it, b));
+            }
 
-    for (auto it = targets_.get()->begin() ; it != targets_.get()->end(); ++it) {
-      std::shared_ptr<arrow::NumericBuilder<TYPE>> b = builders[*it];
-      std::shared_ptr<arrow::Array> array;
-      b->Finish(&array);
-      out.insert(std::pair<int, std::shared_ptr<arrow::Array>>(*it, array));
-    }
-    return 0;
-  }
-};
+            for (int64_t i = 0; i < partitions->length(); i++) {
+                std::shared_ptr<arrow::NumericBuilder<TYPE>> b = builders[partitions->Value(i)];
+                b->Append(reader->Value(i));
+            }
 
-class FixedBinaryArrayMerger : public ArrowArrayMergeKernel {
-public:
-  explicit FixedBinaryArrayMerger(std::shared_ptr<arrow::DataType> type,
-                                        arrow::MemoryPool* pool, std::shared_ptr<std::vector<int>> targets) :
-      ArrowArrayMergeKernel(type, pool, targets) {}
+            for (auto it = targets_.get()->begin(); it != targets_.get()->end(); ++it) {
+                std::shared_ptr<arrow::NumericBuilder<TYPE>> b = builders[*it];
+                std::shared_ptr<arrow::Array> array;
+                b->Finish(&array);
+                out.insert(std::pair<int, std::shared_ptr<arrow::Array>>(*it, array));
+            }
+            return 0;
+        }
+    };
 
-  int Merge(std::shared_ptr<arrow::Array> &values,
-            std::shared_ptr <arrow::Int32Array>& partitions,
-            std::unordered_map<int, std::shared_ptr<arrow::Array>>& out) override;
-};
+    class FixedBinaryArrayMerger : public ArrowArrayMergeKernel {
+    public:
+        explicit FixedBinaryArrayMerger(std::shared_ptr<arrow::DataType> type,
+                                        arrow::MemoryPool *pool, std::shared_ptr<std::vector<int>> targets) :
+                ArrowArrayMergeKernel(type, pool, targets) {}
 
-class BinaryArrayMerger : public ArrowArrayMergeKernel {
-public:
-  explicit BinaryArrayMerger(std::shared_ptr<arrow::DataType> type,
-                                  arrow::MemoryPool* pool, std::shared_ptr<std::vector<int>> targets) :
-      ArrowArrayMergeKernel(type, pool, targets) {}
+        int Merge(std::shared_ptr<arrow::Array> &values,
+                  std::shared_ptr<arrow::Int32Array> &partitions,
+                  std::unordered_map<int, std::shared_ptr<arrow::Array>> &out) override;
+    };
 
-  int Merge(std::shared_ptr<arrow::Array> &values,
-            std::shared_ptr <arrow::Int32Array>& partitions,
-            std::unordered_map<int, std::shared_ptr<arrow::Array>>& out) override;
-};
+    class BinaryArrayMerger : public ArrowArrayMergeKernel {
+    public:
+        explicit BinaryArrayMerger(std::shared_ptr<arrow::DataType> type,
+                                   arrow::MemoryPool *pool, std::shared_ptr<std::vector<int>> targets) :
+                ArrowArrayMergeKernel(type, pool, targets) {}
 
-using UInt8ArrayMerger = ArrowArrayNumericMergeKernel<arrow::UInt8Type>;
-using UInt16ArrayMerger = ArrowArrayNumericMergeKernel<arrow::UInt16Type>;
-using UInt32ArrayMerger = ArrowArrayNumericMergeKernel<arrow::UInt32Type>;
-using UInt64ArrayMerger = ArrowArrayNumericMergeKernel<arrow::UInt64Type>;
+        int Merge(std::shared_ptr<arrow::Array> &values,
+                  std::shared_ptr<arrow::Int32Array> &partitions,
+                  std::unordered_map<int, std::shared_ptr<arrow::Array>> &out) override;
+    };
 
-using Int8ArrayMerger = ArrowArrayNumericMergeKernel<arrow::Int8Type>;
-using Int16ArrayMerger = ArrowArrayNumericMergeKernel<arrow::Int16Type>;
-using Int32ArrayMerger = ArrowArrayNumericMergeKernel<arrow::Int32Type>;
-using Int64ArrayMerger = ArrowArrayNumericMergeKernel<arrow::Int64Type>;
+    using UInt8ArrayMerger = ArrowArrayNumericMergeKernel<arrow::UInt8Type>;
+    using UInt16ArrayMerger = ArrowArrayNumericMergeKernel<arrow::UInt16Type>;
+    using UInt32ArrayMerger = ArrowArrayNumericMergeKernel<arrow::UInt32Type>;
+    using UInt64ArrayMerger = ArrowArrayNumericMergeKernel<arrow::UInt64Type>;
 
-using HalfFloatArrayMerger = ArrowArrayNumericMergeKernel<arrow::HalfFloatType>;
-using FloatArrayMerger = ArrowArrayNumericMergeKernel<arrow::FloatType>;
-using DoubleArrayMerger = ArrowArrayNumericMergeKernel<arrow::DoubleType>;
+    using Int8ArrayMerger = ArrowArrayNumericMergeKernel<arrow::Int8Type>;
+    using Int16ArrayMerger = ArrowArrayNumericMergeKernel<arrow::Int16Type>;
+    using Int32ArrayMerger = ArrowArrayNumericMergeKernel<arrow::Int32Type>;
+    using Int64ArrayMerger = ArrowArrayNumericMergeKernel<arrow::Int64Type>;
 
-int CreateNumericMerge(std::shared_ptr<arrow::DataType>& type,
-                       arrow::MemoryPool* pool, std::shared_ptr<std::vector<int>> targets,
-                       std::unique_ptr<ArrowArrayMergeKernel>* out);
+    using HalfFloatArrayMerger = ArrowArrayNumericMergeKernel<arrow::HalfFloatType>;
+    using FloatArrayMerger = ArrowArrayNumericMergeKernel<arrow::FloatType>;
+    using DoubleArrayMerger = ArrowArrayNumericMergeKernel<arrow::DoubleType>;
 
-class ArrowArraySortKernel {
-public:
-  explicit ArrowArraySortKernel(std::shared_ptr<arrow::DataType> type,
-                                 arrow::MemoryPool* pool) : type_(type), pool_(pool) {}
+    int CreateNumericMerge(std::shared_ptr<arrow::DataType> &type,
+                           arrow::MemoryPool *pool, std::shared_ptr<std::vector<int>> targets,
+                           std::unique_ptr<ArrowArrayMergeKernel> *out);
 
-  /**
-   * Sort the values in the column and return an array with the indices
-   * @param ctx
-   * @param values
-   * @param targets
-   * @param out_length
-   * @param out
-   * @return
-   */
-  virtual int Sort(std::shared_ptr<arrow::Array> values,
-                    std::shared_ptr<arrow::Array>* out) = 0;
-protected:
-  std::shared_ptr<arrow::DataType> type_;
-  arrow::MemoryPool* pool_;
-};
+    class ArrowArraySortKernel {
+    public:
+        explicit ArrowArraySortKernel(std::shared_ptr<arrow::DataType> type,
+                                      arrow::MemoryPool *pool) : type_(type), pool_(pool) {}
 
-template <typename TYPE>
-class ArrowArrayNumericSortKernel : public ArrowArraySortKernel {
-public:
-  using T = typename TYPE::c_type;
+        /**
+         * Sort the values in the column and return an array with the indices
+         * @param ctx
+         * @param values
+         * @param targets
+         * @param out_length
+         * @param out
+         * @return
+         */
+        virtual int Sort(std::shared_ptr<arrow::Array> values,
+                         std::shared_ptr<arrow::Array> *out) = 0;
 
-  explicit ArrowArrayNumericSortKernel(std::shared_ptr<arrow::DataType> type,
-                                        arrow::MemoryPool* pool) :
-      ArrowArraySortKernel(type, pool) {}
+    protected:
+        std::shared_ptr<arrow::DataType> type_;
+        arrow::MemoryPool *pool_;
+    };
 
-  int Sort(std::shared_ptr<arrow::Array> values,
-            std::shared_ptr<arrow::Array>* offsets) override {
-    auto array = std::static_pointer_cast<arrow::NumericArray<TYPE>>(values);
-    const T* left_data = array->raw_values();
-    std::shared_ptr<arrow::Buffer> indices_buf;
-    int64_t buf_size = values->length() * sizeof(uint64_t);
-    arrow::Status status = AllocateBuffer(arrow::default_memory_pool(), buf_size + 1, &indices_buf);
-    if (status != arrow::Status::OK()) {
-      //LOG(FATAL) << "Failed to allocate sort indices - " << status.message();
-      return -1;
-    }
-    auto *indices_begin = reinterpret_cast<int64_t *>(indices_buf->mutable_data());
-    for (int64_t i = 0; i < values->length(); i++) {
-      indices_begin[i] = i;
-    }
-    int64_t *indices_end = indices_begin + values->length();
-    // LOG(INFO) << "Length " << values->length() << " ind " << indices_begin << "nd " << indices_begin + values->length();
-    std::sort(indices_begin, indices_end, [left_data](uint64_t left, uint64_t right) {
-      return left_data[left] < left_data[right];
-    });
-    *offsets = std::make_shared<arrow::UInt64Array>(values->length(), indices_buf);
-    return 0;
-  }
-};
+    template<typename TYPE>
+    class ArrowArrayNumericSortKernel : public ArrowArraySortKernel {
+    public:
+        using T = typename TYPE::c_type;
+
+        explicit ArrowArrayNumericSortKernel(std::shared_ptr<arrow::DataType> type,
+                                             arrow::MemoryPool *pool) :
+                ArrowArraySortKernel(type, pool) {}
+
+        int Sort(std::shared_ptr<arrow::Array> values,
+                 std::shared_ptr<arrow::Array> *offsets) override {
+            auto array = std::static_pointer_cast<arrow::NumericArray<TYPE>>(values);
+            const T *left_data = array->raw_values();
+            std::shared_ptr<arrow::Buffer> indices_buf;
+            int64_t buf_size = values->length() * sizeof(uint64_t);
+            arrow::Status status = AllocateBuffer(arrow::default_memory_pool(), buf_size + 1, &indices_buf);
+
+            if (status.OK().code() != arrow::Status::OK().code()) {
+                //LOG(FATAL) << "Failed to allocate sort indices - " << status.message();
+                return -1;
+            }
+            auto *indices_begin = reinterpret_cast<int64_t *>(indices_buf->mutable_data());
+            for (int64_t i = 0; i < values->length(); i++) {
+                indices_begin[i] = i;
+            }
+            int64_t *indices_end = indices_begin + values->length();
+            // LOG(INFO) << "Length " << values->length() << " ind " << indices_begin << "nd " << indices_begin + values->length();
+            std::sort(indices_begin, indices_end, [left_data](uint64_t left, uint64_t right) {
+                return left_data[left] < left_data[right];
+            });
+            *offsets = std::make_shared<arrow::UInt64Array>(values->length(), indices_buf);
+            return 0;
+        }
+    };
 
 
+    using UInt8ArraySorter = ArrowArrayNumericSortKernel<arrow::UInt8Type>;
+    using UInt16ArraySorter = ArrowArrayNumericSortKernel<arrow::UInt16Type>;
+    using UInt32ArraySorter = ArrowArrayNumericSortKernel<arrow::UInt32Type>;
+    using UInt64ArraySorter = ArrowArrayNumericSortKernel<arrow::UInt64Type>;
+    using Int8ArraySorter = ArrowArrayNumericSortKernel<arrow::Int8Type>;
+    using Int16ArraySorter = ArrowArrayNumericSortKernel<arrow::Int16Type>;
+    using Int32ArraySorter = ArrowArrayNumericSortKernel<arrow::Int32Type>;
+    using Int64ArraySorter = ArrowArrayNumericSortKernel<arrow::Int64Type>;
+    using HalfFloatArraySorter = ArrowArrayNumericSortKernel<arrow::HalfFloatType>;
+    using FloatArraySorter = ArrowArrayNumericSortKernel<arrow::FloatType>;
+    using DoubleArraySorter = ArrowArrayNumericSortKernel<arrow::DoubleType>;
 
-using UInt8ArraySorter = ArrowArrayNumericSortKernel<arrow::UInt8Type>;
-using UInt16ArraySorter = ArrowArrayNumericSortKernel<arrow::UInt16Type>;
-using UInt32ArraySorter = ArrowArrayNumericSortKernel<arrow::UInt32Type>;
-using UInt64ArraySorter = ArrowArrayNumericSortKernel<arrow::UInt64Type>;
-using Int8ArraySorter = ArrowArrayNumericSortKernel<arrow::Int8Type>;
-using Int16ArraySorter = ArrowArrayNumericSortKernel<arrow::Int16Type>;
-using Int32ArraySorter = ArrowArrayNumericSortKernel<arrow::Int32Type>;
-using Int64ArraySorter = ArrowArrayNumericSortKernel<arrow::Int64Type>;
-using HalfFloatArraySorter = ArrowArrayNumericSortKernel<arrow::HalfFloatType>;
-using FloatArraySorter = ArrowArrayNumericSortKernel<arrow::FloatType>;
-using DoubleArraySorter = ArrowArrayNumericSortKernel<arrow::DoubleType>;
-
-arrow::Status SortIndices(arrow::MemoryPool *memory_pool, std::shared_ptr<arrow::Array> values,
-                              std::shared_ptr<arrow::Array>* offsets);
+    arrow::Status SortIndices(arrow::MemoryPool *memory_pool, std::shared_ptr<arrow::Array> values,
+                              std::shared_ptr<arrow::Array> *offsets);
 
 }
